requirements.txt
fastapi[all]
fastapi_cache2[redis]
fastapi-utils
uvicorn[standard]
gunicorn[standard]
firebase-admin
aioredis
aiofiles
asyncio
boto3